requirements.txt
flask==2.2.3
opencv-python==4.6.0.66
numpy==1.23.5
pillow==8.4.0
pytest==8.3.3
